move release notes out of Chat
