+++ components/Chat/Chat.tsx
@@ -9,9 +9,7 @@ import {
 } from "@tabler/icons-react"
 import React, {MutableRefObject, memo, useCallback, useContext, useEffect, useRef, useState} from "react"
 import toast from "react-hot-toast"
-import Modal from "react-modal"
 import {useTranslation} from "next-i18next"
-import {useRouter} from "next/router"
 import {useFetch} from "@/hooks/useFetch"
 import useApiService from "@/services/useApiService"
 import {RESPONSE_TIMEOUT_MS, TOAST_DURATION_MS} from "@/utils/app/const"
@@ -22,8 +20,8 @@ import {ChatBody, Conversation, Message} from "@/types/chat"
 import {fallbackOpenAIModel} from "@/types/openai"
 import {Plugin} from "@/types/plugin"
 import HomeContext from "@/pages/api/home/home.context"
+import ReleaseNotes from "@/components/Chat/ReleaseNotes"
 import {WelcomeMessage} from "@/components/Chat/WelcomeMessage"
-import {MemoizedReactMarkdown} from "@/components/Markdown/MemoizedReactMarkdown"
 import Spinner from "../Spinner"
 import {ChatInput} from "./ChatInput"
 import {ChatLoader} from "./ChatLoader"
@@ -31,26 +29,11 @@ import {ErrorMessageDiv} from "./ErrorMessageDiv"
 import {MemoizedChatMessage} from "./MemoizedChatMessage"
 import {ModelSelect} from "./ModelSelect"
 import {toPng} from "html-to-image"
-import rehypeMathjax from "rehype-mathjax"
-import remarkGfm from "remark-gfm"
-import remarkMath from "remark-math";
 
 
 interface Props {
   stopConversationRef: MutableRefObject<boolean>
   theme: "light" | "dark"
-}
-
-const useMarkdownFile = (filename: string) => {
-  const [fileContent, setFileContent] = useState<string | null>(null)
-
-  useEffect(() => {
-    fetch(filename)
-      .then((response) => response.text())
-      .then((text) => setFileContent(text))
-      .catch((error) => console.error(`Error fetching markdown file: ${error}`))
-  }, [filename])
-  return fileContent
 }
 
 export const Chat = memo(({stopConversationRef, theme}: Props) => {
@@ -78,34 +61,12 @@ export const Chat = memo(({stopConversationRef, theme}: Props) => {
   const [showSettings, setShowSettings] = useState<boolean>(false)
   const [showScrollDownButton, setShowScrollDownButton] = useState<boolean>(false)
   const [isReleaseNotesDialogOpen, setIsReleaseNotesDialogOpen] = useState<boolean>(false)
-  const router = useRouter()
   const {getEndpoint} = useApiService()
   const fetchService = useFetch()
 
-  const releaseNotesMarkdown = useMarkdownFile(`${router.basePath}/RELEASE_NOTES.md`)
   const messagesEndRef = useRef<HTMLDivElement>(null)
   const chatContainerRef = useRef<HTMLDivElement>(null)
   const textareaRef = useRef<HTMLTextAreaElement>(null)
-
-  const customModalStyles = {
-    content: {
-      backgroundColor: "#e6e6e0",
-      color: "#000000",
-      top: "50%",
-      left: "50%",
-      right: "auto",
-      bottom: "auto",
-      marginRight: "-50%",
-      transform: "translate(-50%, -50%)",
-      padding: "2rem",
-      maxWidth: "50%",
-      maxHeight: "80%",
-      overflow: "auto"
-    },
-    overlay: {
-      backgroundColor: "rgba(0, 0, 0, 0.5)"
-    }
-  }
 
   const handleSend = useCallback(
     async (message: Message, deleteCount = 0, plugin: Plugin | null = null) => {
@@ -570,51 +531,7 @@ export const Chat = memo(({stopConversationRef, theme}: Props) => {
           )}
         </>
       )}
-      <Modal
-        isOpen={isReleaseNotesDialogOpen}
-        onRequestClose={() => setIsReleaseNotesDialogOpen(false)}
-        style={customModalStyles}
-        contentLabel="Release Notes"
-        ariaHideApp={false}
-      >
-        <MemoizedReactMarkdown
-          className="prose dark:prose-invert flex-1"
-          remarkPlugins={[remarkGfm, remarkMath]}
-          rehypePlugins={[rehypeMathjax]}
-          components={{
-            code({node, inline, className, children, ...props}) {
-              return (
-                <code className={className} {...props}>
-                  {children}
-                </code>
-              )
-            },
-            table({children}) {
-              return (
-                <table className="border-collapse border border-black px-3 py-1 dark:border-white">{children}</table>
-              )
-            },
-            th({children}) {
-              return (
-                <th className="break-words border border-black bg-gray-500 px-3 py-1 text-white dark:border-white">
-                  {children}
-                </th>
-              )
-            },
-            td({children}) {
-              return <td className="break-words border border-black px-3 py-1 dark:border-white">{children}</td>
-            }
-          }}
-        >
-          {`${releaseNotesMarkdown ? releaseNotesMarkdown : `Loading release notes...`}`}
-        </MemoizedReactMarkdown>
-        <button
-          className="h-[40px] rounded-md bg-blue-500 px-4 py-1 text-sm font-medium text-white enabled:hover:bg-blue-600 disabled:opacity-50"
-          onClick={() => setIsReleaseNotesDialogOpen(false)}
-        >
-          Dismiss
-        </button>
-      </Modal>
+      <ReleaseNotes isOpen={isReleaseNotesDialogOpen} close={() => setIsReleaseNotesDialogOpen(false)} />
     </div>
   )
 })
